crap/backend/requirements.txt
supabase==0.7.1
python-dotenv==0.19.0
